requirements.txt
requests>=2.31.0
python-jose[cryptography]>=3.3.0
SQLAlchemy>=2.0.0

# Drivers adicionais — descomente conforme seu banco:
# pyodbc>=5.0.0          # SQL Server
# psycopg2-binary>=2.9   # PostgreSQL
